Refactor and cleanup image preview modal components
