+++ src/components/image_preview_modal/ImagePreviewModal.tsx
@@ -1,31 +1,13 @@
-/**
- * Author : Amit raikwar
- * Last updated : 06 Feb, 2024
- */
-//~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
+import { IMAGE_MODAL_ID } from "./constants";
+import { ImagePreviewModalProps } from "./types";
 
-/**
- * The image modal component id.
- */
-const IMAGE_MODAL_ID = "modal_image";
-
-//~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
-/**
- * The image modal props.
- * @param image The image url to be displayed in the modal.
- */
-interface IImagePreviewModalProps {
-  image: string;
-}
-
-//~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
 /**
  * Image modal component.
  *
  * @param props  The image modal props.
  * @returns The image modal component.
  */
-function ImagePreviewModal(props: IImagePreviewModalProps) {
+export const ImagePreviewModal = ({ image }: ImagePreviewModalProps) => {
   return (
     <div className="z-50 ">
       <input type="checkbox" id={IMAGE_MODAL_ID} className="modal-toggle" />
@@ -40,7 +22,7 @@ function ImagePreviewModal(props: IImagePreviewModalProps) {
             </label>
           </div>
           <img
-            src={props.image}
+            src={image}
             alt="preview"
             loading="lazy"
             className=" rounded-xl object-cover w-full h-[90%]"
@@ -49,14 +31,4 @@ function ImagePreviewModal(props: IImagePreviewModalProps) {
       </div>
     </div>
   );
-}
-
-//~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
-// Export the ImagePreviewModal component.
-export default ImagePreviewModal;
-export type { IImagePreviewModalProps };
-export { IMAGE_MODAL_ID };
-
-//~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
-// End of file.
-//~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
+};
