+++ src/components/image_preview_modal/constants.ts
@@ -1,0 +1,4 @@
+/**
+ * The image modal component id.
+ */
+export const IMAGE_MODAL_ID = "modal_image";
