+++ src/components/image_preview_modal/index.ts
@@ -1,6 +1,7 @@
-export { default as ImagePreviewModalButton } from "./ImagePreviewButton";
-export type { IImagePreviewModalButtonProps } from "./ImagePreviewButton";
-export { IMAGE_MODAL_ID } from "./ImagePreviewModal";
-
-export { default as ImagePreviewModal } from "./ImagePreviewModal";
-export type { IImagePreviewModalProps } from "./ImagePreviewModal";
+export { ImagePreviewModalButton } from "./ImagePreviewButton";
+export { ImagePreviewModal } from "./ImagePreviewModal";
+export type {
+  ImagePreviewModalButtonProps,
+  ImagePreviewModalProps,
+} from "./types";
+export { IMAGE_MODAL_ID } from "./constants";
