+++ src/components/image_preview_modal/types.ts
@@ -1,0 +1,8 @@
+export type ImagePreviewModalProps = {
+  image: string;
+};
+
+export type ImagePreviewModalButtonProps = {
+  onClickHandler: () => void;
+  children: React.ReactNode;
+};
